modifiedc

+++ 7thfeb/lab18Task.py
@@ -1,5 +1,5 @@
 import math
-a = int(input("Enter the radius of the circle\n"))
+a = float(input("Enter the radius of the circle\n"))
 Area = math.pi * pow(a, 2)
 print(Area)
